local debugged; demo; datatype unsigned long long

+++ plc_assign3/functions.h
@@ -1,6 +1,6 @@
 #include <mpi.h>
 
 
-int MPI_P2P_reduce(const int *sendbuf, int *recvbuf, int count,
+int MPI_P2P_reduce(const unsigned long long *sendbuf, unsigned long long *recvbuf, int count,
                        MPI_Datatype datatype, MPI_Op op, int root,
                        MPI_Comm comm);
